backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pymysql==1.1.2
python-dotenv==1.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.4.0
pydantic==2.11.1
requests>=2.31.0
scikit-learn>=1.4.0
pandas>=2.1.0
httpx>=0.27.0
pytest>=8.0.0
pyotp>=2.9.0
